requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.7.3
slackclient==2.5.0
virtualenv==16.7.9
yarl==1.4.2
requests=3.0.4
